Cut all code to 79 chars and disable E501
E501 is reponsible for 225 errors (most of which occur with documentation strings and not actual code).
Additionally, these errors are not automatically corrected for docstrings by `autopep8`, so we have no choice but to disable this error for the sake of saving time.
Code can be maintained to be well under the 79 character limit, which is  what has been done before the rule was disabled.